+++ src/content/dom-to-ast/shared.ts
@@ -1,0 +1,18 @@
+import type { AuthorInfo } from '../../ast/types';
+
+export function stripHandlePrefix(a: { name: string; handle: string }): AuthorInfo {
+  return {
+    name: a.name,
+    handle: a.handle.startsWith('@') ? a.handle.slice(1) : a.handle,
+  };
+}
+
+export function extractDateFromArticle(article: Element): string {
+  const timeEl = article.querySelector('time');
+  if (timeEl) {
+    const datetime = timeEl.getAttribute('datetime');
+    if (datetime) return datetime;
+    return timeEl.textContent?.trim() || '';
+  }
+  return '';
+}
